+++ docs/CHANGELOG.md
@@ -5,3 +5,5 @@
 - **2025-08-27**
   - Refactor user validation in `checker.py` to return warnings alongside errors for logging.
   - Refactor keyword validation in `checker.py`
+  - Add output format to SERIF, JSON or TEXT `main.py --format ["text", "json", "sarif"]`
+  - Update `README.md`